requirements.txt
Flask
Flask-CORS
python-dotenv
gunicorn
PyMuPDF
google-generativeai
